requirements.txt
requests==2.22.0
pillow==7.1.1
python-dotenv==0.12.0
instabot==0.117.0
